requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
pydantic[email]
bcrypt
python-jose[cryptography]
python-multipart
llama-parse
langchain
langchain-text-splitters
chromadb
sentence-transformers
torch